apps/core/api/viewsets.py
# -*- coding: utf-8 -*-

__author__ = 'Nikolay Mamashin (mamashin@gmail.com)'

from django.shortcuts import render
from django.conf import settings
from rest_framework import viewsets
from rest_framework.views import APIView
from rest_framework.response import Response
from django.utils import timezone
from django.http import HttpResponse
from django.core.exceptions import PermissionDenied
from decouple import config
from rest_framework.permissions import IsAuthenticated, IsAuthenticatedOrReadOnly, AllowAny
from apps.core.api.permissions import SuperuserAccessPermission, OnlySuperuserRW
import time
import re
from pathlib import Path
from django.db.models import Value, BooleanField
