pyproject.toml
[tool.poetry]
name = "epc-qr-generator"
version = "0.1.0"
description = "EPC QR Generator"
authors = ["mamashin <mamashin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.0"
loguru = "^0.7.2"
python-decouple = "^3.5"
djangorestframework = "^3.13.1"
httpx = "^0.27.0"
sentry-sdk = "^1.5.2"
pillow = "^10.1.0"
django-behaviors = "^0.5.1"
django-split-settings = "^1.2.0"
django-ipware = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
